把简历,收藏简历和成员表增加学号一列, 更改了登陆方法为学号加密码, 测试各个API已通过

+++ src/main/java/com/turing/turing/dao/CollectResumeMapper.java
@@ -2,11 +2,12 @@ package com.turing.turing.dao;
 
 import com.turing.turing.entity.CollectResume;
 import com.turing.turing.entity.CollectResumeExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
 public interface CollectResumeMapper {
-    long countByExample(CollectResumeExample example);
+    int countByExample(CollectResumeExample example);
 
     int deleteByExample(CollectResumeExample example);
