+++ src/main/java/com/turing/turing/dao/MemberMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface MemberMapper {
-    long countByExample(MemberExample example);
+    int countByExample(MemberExample example);
 
     int deleteByExample(MemberExample example);
