+++ src/main/java/com/turing/turing/dao/ResumeMapper.java
@@ -2,11 +2,12 @@ package com.turing.turing.dao;
 
 import com.turing.turing.entity.Resume;
 import com.turing.turing.entity.ResumeExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
 public interface ResumeMapper {
-    long countByExample(ResumeExample example);
+    int countByExample(ResumeExample example);
 
     int deleteByExample(ResumeExample example);
